# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server.",
        "title": "My Golang API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/api/auth": {
            "get": {
                "description": "Получение информации о пользователе по идентификатору",
                "tags": [
                    "Пользователи"
                ],
                "summary": "Возвращает информацию о пользователе по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID пользователя",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "Bad"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "Not"
                        }
                    }
                }
            },
            "post": {
                "description": "Создание нового пользователя",
                "tags": [
                    "Пользователи"
                ],
                "summary": "Создание нового пользователя",
                "parameters": [
                    {
                        "description": "Данные для создания пользователя",
                        "name": "createUser",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "type": "Unprocessable"
                        }
                    }
                }
            }
        },
...
        "/api/users": {
            "get": {
                "description": "Получение списка всех пользователей",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Пользователи"
                ],
                "summary": "Возвращает список всех пользователей",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.User"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "Bad"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "Internal"
                        }
                    }
                }
            },
            "put": {
                "description": "Обновление существующего пользователя",
                "tags": [
                    "Пользователи"
                ],
                "summary": "Обновление существующего пользователя",
                "parameters": [
                    {
                        "description": "Данные для обновления пользователя",
                        "name": "updateUser",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateUserRequest"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "ID пользователя",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "type": "No"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "Not"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "type": "Unprocessable"
                        }
                    }
                }
            }
        },
        "/api/users/{id}": {
            "get": {
                "description": "Получение информации о пользователе по идентификатору",
                "tags": [
                    "Пользователи"
                ],
                "summary": "Возвращает информацию о пользователе по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID пользователя",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "Bad"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "Not"
                        }
                    }
                }
            },
            "delete": {
                "description": "Удаление пользователя по идентификатору",
                "tags": [
                    "Пользователи"
                ],
                "summary": "Удаляет пользователя по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID пользователя",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "type": "No"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "Not"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.CreateUserRequest": {
            "type": "object",
            "required": [
                "email",
                "name",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string",
                    "maxLength": 128,
                    "minLength": 3
                },
                "password": {
                    "type": "string",
                    "maxLength": 128,
                    "minLength": 4
                },
                "phone": {
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 6
                }
            }
        },
...
        "models.UpdateUserRequest": {
            "type": "object",
            "required": [
                "email",
                "name",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string",
                    "maxLength": 128,
                    "minLength": 3
                },
                "password": {
                    "type": "string",
                    "maxLength": 128,
                    "minLength": 4
                },
                "phone": {
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 6
                }
            }
        },
        "models.User": {
            "type": "object",
            "properties": {
                "city": {
                    "description": "Город",
                    "type": "string"
                },
                "created_at": {
                    "description": "Дата создания",
                    "type": "string"
                },
                "deleted_at": {
                    "description": "Дата удаления",
                    "type": "string"
                },
                "email": {
                    "description": "Email",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "logo": {
                    "description": "Логотип",
                    "type": "string"
                },
                "media": {
                    "description": "Медиа",
                    "type": "string"
                },
                "name": {
                    "description": "ФИО",
                    "type": "string"
                },
                "password": {
                    "description": "Пароль",
                    "type": "string"
                },
                "phone": {
                    "description": "Телефон",
                    "type": "string"
                },
                "status": {
                    "description": "Статус",
                    "type": "integer"
                },
                "updated_at": {
                    "description": "Дата последнего обновления",
                    "type": "string"
                }
            }
        },
...
    }
}